move types to types folder simplify npm scripts
