+++ .taprc
@@ -1,6 +1,2 @@
-100: true
-ts: false
-jsx: false
-flow: false
-coverage: true
-check-coverage: true
+files:
+  - 'test/**/*.test.js'
